+++ src/components/Select/Select.types.ts
@@ -29,7 +29,6 @@ export interface SelectProps<T extends object> extends Omit<
   children: ReactNode | ((item: T) => ReactNode);
   slots?: SelectSlots<T>;
   popoverOffset?: number;
-  popoverPortalContainer?: () => HTMLElement | null;
   classNames?: SlotsToClasses<
     | 'label'
     | 'name'